# uftos-openapi-specs.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "UFTOS OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost:5173/api",
    "description" : "UFTOS api URL"
  } ],
  "paths" : {
...
    "/subjects" : {
      "delete" : {
        "operationId" : "deleteSubjects",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "subject-controller" ]
      },
      "get" : {
        "operationId" : "getSubjects",
        "parameters" : [ {
          "in" : "query",
          "name" : "sort",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Sort"
          }
        }, {
          "in" : "query",
          "name" : "search",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Subject"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "subject-controller" ]
      },
      "post" : {
        "operationId" : "createSubject",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SubjectRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Subject"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "subject-controller" ]
      }
    },
    "/subjects/{id}" : {
      "delete" : {
        "operationId" : "deleteSubject",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "subject-controller" ]
      },
      "get" : {
        "operationId" : "getSubject",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Subject"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "subject-controller" ]
      },
      "put" : {
        "operationId" : "updateSubject",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SubjectRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Subject"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "subject-controller" ]
      }
    },
    "/tags" : {
      "delete" : {
        "operationId" : "deleteTags",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "tag-controller" ]
      },
      "get" : {
        "operationId" : "getTags",
        "parameters" : [ {
          "in" : "query",
          "name" : "sort",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Sort"
          }
        }, {
          "in" : "query",
          "name" : "search",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Tag"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "tag-controller" ]
      },
      "post" : {
        "operationId" : "createTag",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TagRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Tag"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "tag-controller" ]
      }
    },
    "/tags/{id}" : {
      "delete" : {
        "operationId" : "deleteTag",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "tag-controller" ]
      },
      "get" : {
        "operationId" : "getTag",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Tag"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "tag-controller" ]
      },
      "put" : {
        "operationId" : "updateTag",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TagRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Tag"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "tag-controller" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Sort" : {
        "type" : "object",
        "properties" : {
          "sort" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "Subject" : {
        "required" : [ "id", "name", "tags" ],
        "type" : "object",
        "properties" : {
          "color" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Tag"
            }
          }
        }
      },
      "SubjectRequestDto" : {
        "required" : [ "name", "tagIds" ],
        "type" : "object",
        "properties" : {
          "color" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "tagIds" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "Tag" : {
        "required" : [ "id", "name" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
      "TagRequestDto" : {
        "required" : [ "tagName" ],
        "type" : "object",
        "properties" : {
          "tagName" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}